Replace theme toggle with Classic React toggle
